requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
pdfplumber==0.10.3
python-docx==1.1.0
langchain==0.1.0
langchain-google-genai==0.0.6